mywebsite/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-ckeditor==6.3.2
django-js-asset==2.0.0
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
